+++ internal/db/sqltest/tests/pagination/session_recording.sql
@@ -4,8 +4,16 @@
 begin;
   select plan(2);
 
-  select has_index('recording_session', 'recording_session_create_time_public_id_idx', array['create_time', 'public_id']);
-  select has_index('recording_session', 'recording_session_update_time_public_id_idx', array['update_time', 'public_id']);
+  select has_index(
+    'recording_session',
+    'recording_session_create_time_public_id_delete_time_delete_idx',
+    array['create_time', 'public_id', 'delete_time', 'delete_after']
+  );
+  select has_index(
+    'recording_session',
+    'recording_session_update_time_public_id_delete_time_delete_idx',
+    array['update_time', 'public_id', 'delete_time', 'delete_after']
+  );
 
   select * from finish();